treewide: implement xdg-toplevel-getinfo
this executable will return the current height and width available for any window to draw on. we want to call this during android startup.
it will set squeekboard to invisible and return a good value. different devices can implement this differently if they lets say want to always return one value

Signed-off-by: Bardia Moshiri <bardia@furilabs.com>